Investigue la publication du paquet npm publicode
