+++ .gitignore
@@ -5,6 +5,7 @@ yarn-error.log
 package-lock.json
 node_modules/
 .env
+dist/
 
 # Local Netlify folder
 .netlify
