+++ babel.config.json
@@ -17,7 +17,6 @@
 		"@babel/plugin-proposal-optional-chaining",
 		"@babel/plugin-proposal-nullish-coalescing-operator",
 		"@babel/plugin-proposal-object-rest-spread",
-		"@babel/plugin-syntax-dynamic-import",
-		["ramda", { "useES": true }]
+		"@babel/plugin-syntax-dynamic-import"
 	]
 }
