+++ mon-entreprise/.babelrc.json
@@ -1,6 +1,7 @@
 {
 	"plugins": [
 		"react-hot-loader/babel",
-		["webpack-alias", { "config": "./webpack.dev.js" }]
+		["webpack-alias", { "config": "./webpack.dev.js" }],
+		["ramda", { "useES": true }]
 	]
 }
