+++ mon-entreprise/.gitignore
@@ -1,8 +1,4 @@
 .env
-node_modules/
-package-lock.json
-yarn-error.log
 source/data
-dist/
 cypress/videos
 cypress/screenshots
