+++ mon-entreprise/package.json
@@ -16,7 +16,12 @@
 		"not ie < 11"
 	],
 	"devDependencies": {
-		"i18next-parser": "https://github.com/i18next/i18next-parser#master"
+		"copy-webpack-plugin": "^4.5.2",
+		"html-webpack-plugin": "^3.2.0",
+		"i18next-parser": "^1.0.6",
+		"monaco-editor-webpack-plugin": "^1.9.0",
+		"workbox-webpack-plugin": "^5.1.3",
+		"worker-loader": "^2.0.0"
 	},
 	"dependencies": {
 		"@babel/runtime": "^7.3.4",
@@ -31,8 +36,7 @@
 		"js-yaml": "^3.13.1",
 		"moo": "^0.5.0",
 		"nearley": "^2.19.0",
-		"publicodes": "file:../publicodes",
-		"puppeteer": "^2.1.1",
+		"publicodes": "^1.0.0",
 		"ramda": "^0.27.0",
 		"react": "^16.13.1",
 		"react-color": "^2.14.0",
@@ -72,7 +76,7 @@
 		"compile:analyze-bundle": "ANALYZE_BUNDLE=1 yarn run compile",
 		"compile:dev": "FR_SITE='http://localhost:5000${path}' EN_SITE='http://localhost:5001${path}' yarn run compile",
 		"test": "yarn test:file \"./{,!(node_modules)/**/}!(webpack).test.js\"",
-		"test:file": "yarn mocha-webpack --webpack-config ../webpack.test.js --require source-map-support/register --include test/componentTestSetup.js --require mock-local-storage --require test/helpers/browser.js",
+		"test:file": "yarn mocha-webpack --webpack-config ./webpack.dev.js --include test/componentTestSetup.js --require mock-local-storage --require test/helpers/browser.js",
 		"test:bundlesize": "bundlesize",
 		"test:dev-e2e:publicode": "cypress open --browser chromium --config baseUrl=http://localhost:8080/publicodes,integrationFolder=cypress/integration/publi.codes",
 		"test:dev-e2e:mon-entreprise": "cypress open --browser chromium",
