+++ mon-entreprise/webpack.prod.legacyBrowser.js
@@ -14,7 +14,8 @@ module.exports = {
 	...prod,
 	entry: map(entry => ['whatwg-fetch', entry], prod.entry),
 	output: {
-		filename: '[name].legacy.bundle.js'
+		filename: '[name].legacy.bundle.js',
+		ecmaVersion: 5
 	},
 	module: {
 		rules: [...commonLoaders({ legacy: true }), styleLoader('style-loader')]
