+++ netlify.toml
@@ -1,65 +1,65 @@
 # Redirects following architectural changes on the end of October 2018
-[[redirects]] 
+[[redirects]]
   from="/s%C3%A9curit%C3%A9-sociale"
   to="/g%C3%A9rer"
   status = 301
 
-[[redirects]] 
+[[redirects]]
   from="/d%C3%A9marches-embauche"
   to="/g%C3%A9rer/embaucher"
   status = 301
 
-[[redirects]] 
+[[redirects]]
   from="/s%C3%A9curit%C3%A9-sociale/s%C3%A9lection-du-r%C3%A9gime"
   to="/simulateurs"
   status = 301
 
-[[redirects]] 
+[[redirects]]
   from="/s%C3%A9curit%C3%A9-sociale/*"
   to="/simulateurs/:splat"
   status = 301
 
-[[redirects]] 
+[[redirects]]
   from="/entreprise/cr%C3%A9er-une-*"
   to="/cr%C3%A9er/:splat"
   status = 301
 
-[[redirects]] 
+[[redirects]]
   from="/entreprise/devenir-*"
   to="/cr%C3%A9er/:splat"
   status = 301
 
-[[redirects]] 
+[[redirects]]
   from="/entreprise/*"
   to="/cr%C3%A9er/:splat"
   status = 301
 
-[[redirects]] 
+[[redirects]]
   from="/social-security"
   to="/manage"
   status = 301
 
-[[redirects]] 
+[[redirects]]
   from="/social-security/assimilated-salaried"
   to="/social-security/assimile-salarie"
   status = 301
 
-[[redirects]] 
+[[redirects]]
   from="/social-security/*"
   to="/simulators/:splat"
   status = 301
 
-[[redirects]] 
+[[redirects]]
   from="/company/create-a-*"
   to="/create/:splat"
   status = 301
 
-[[redirects]] 
+[[redirects]]
   from="/company/become-*"
   to="/create/:splat"
   status = 301
 
-[[redirects]] 
+[[redirects]]
   from="/company/*"
   to="/create/:splat"
   status = 301
