+++ package.json
@@ -21,8 +21,6 @@
 		"@types/cheerio": "^0.22.18",
 		"@types/classnames": "^2.2.9",
 		"@types/color-convert": "^1.9.0",
-		"@types/copy-webpack-plugin": "^5.0.0",
-		"@types/html-webpack-plugin": "^3.2.2",
 		"@types/iframe-resizer": "^3.5.7",
 		"@types/js-yaml": "^3.12.2",
 		"@types/mini-css-extract-plugin": "^0.9.1",
@@ -42,7 +40,6 @@
 		"@types/webpack": "^4.41.10",
 		"@types/webpack-bundle-analyzer": "^2.13.3",
 		"@types/webpack-env": "^1.14.1",
-		"@types/workbox-webpack-plugin": "^5.1.0",
 		"@typescript-eslint/eslint-plugin": "^2.31.0",
 		"@typescript-eslint/parser": "^2.31.0",
 		"autoprefixer": "^9.7.6",
@@ -55,7 +52,6 @@
 		"bundlesize": "^0.18.0",
 		"chai": "^4.1.2",
 		"connect-history-api-fallback": "^1.5.0",
-		"copy-webpack-plugin": "^4.5.2",
 		"css-loader": "^0.28.11",
 		"csv-loader": "^2.1.1",
 		"dedent-js": "^1.0.1",
@@ -69,7 +65,6 @@
 		"express": "^4.16.3",
 		"file-loader": "^1.1.11",
 		"html-loader": "^0.5.5",
-		"html-webpack-plugin": "^3.2.0",
 		"http-server": "^0.11.1",
 		"intl": "^1.2.5",
 		"intl-locales-supported": "^1.0.0",
@@ -92,18 +87,15 @@
 		"serve": "^11.1.0",
 		"sinon": "^4.5.0",
 		"sinon-chai": "^3.0.0",
-		"source-map-support": "^0.5.4",
 		"style-loader": "^0.23.1",
 		"toml-loader": "^1.0.0",
 		"typescript": "^3.9.2",
 		"url-loader": "^1.0.1",
-		"webpack": "^4.39.3",
+		"webpack": "^4.42.0",
 		"webpack-bundle-analyzer": "^3.7.0",
 		"webpack-cli": "^3.1.2",
 		"webpack-dev-middleware": "^3.4.0",
 		"webpack-hot-middleware": "^2.24.2",
-		"workbox-webpack-plugin": "^3.6.1",
-		"worker-loader": "^2.0.0",
 		"yaml-jest": "^1.0.5",
 		"yaml-loader": "^0.5.0"
 	},
@@ -113,17 +105,18 @@
 	},
 	"scripts": {
 		"lint:eslintrc": "eslint --print-config .eslintrc.js | eslint-config-prettier-check",
-		"lint:eslint": "eslint . --ext .js,.jsx,.ts,.tsx",
+		"lint:eslint": "export NODE_OPTIONS='--max-old-space-size=4096'; eslint . --ext .js,.jsx,.ts,.tsx",
 		"lint:eslint:fix": "yarn lint:eslint --fix",
 		"lint:prettier": "yarn workspaces run prettier --check \"{source,test,cypress}/**/*.{js,jsx,ts,tsx}\"",
 		"lint:prettier:fix": "yarn lint:prettier --write",
 		"lint:fix": "yarn lint:eslint:fix && yarn lint:prettier:fix",
 		"lint": "yarn lint:eslintrc && yarn lint:eslint && yarn lint:prettier",
 		"test": "yarn workspaces run test",
-		"prepare": "yarn workspace mon-entreprise run prepare",
+		"prepare": "yarn link:publicodes && yarn workspaces run prepare",
 		"test:type": "yarn workspaces run tsc",
 		"test:regressions": "jest",
-		"link:publicodes": "yarn workspace publicodes link && yarn workspace mon-entreprise link publicodes"
+		"link:publicodes": "yarn workspace publicodes link && yarn workspace mon-entreprise link publicodes",
+		"start": "yarn workspace publicodes compile --watch & yarn workspace mon-entreprise start"
 	},
 	"workspaces": [
 		"mon-entreprise",
