+++ publicodes/source/components/rule/Rule.tsx
@@ -2,7 +2,6 @@ import React, { useState } from 'react'
 import { Trans } from 'react-i18next'
 import Engine from '../..'
 import { formatValue } from '../../format'
-import mecanisms from '../../mecanisms.yaml'
 import { serializeUnit } from '../../units'
 import { Markdown } from '../Markdown'
 import { RuleLinkWithContext } from '../RuleLink'
