+++ publicodes/source/types/import-markdown.d.ts
@@ -1,0 +1,4 @@
+declare module '*.md' {
+	const content: string
+	export default content
+}
