+++ publicodes/source/types/import-nearley.d.ts
@@ -1,0 +1,4 @@
+declare module '*.ne' {
+	const content: any
+	export default content
+}
