+++ publicodes/webpack.test.js
@@ -1,0 +1,21 @@
+const { commonLoaders } = require('../webpack/common')
+
+module.exports = {
+	resolve: {
+		extensions: ['.ts', '.tsx', '.js']
+	},
+	output: {
+		devtoolFallbackModuleFilenameTemplate: '[absolute-resource-path]?[hash]',
+		devtoolModuleFilenameTemplate: '[absolute-resource-path]'
+	},
+	mode: 'development',
+	module: {
+		rules: [
+			...commonLoaders(),
+			{
+				test: /\.css$/,
+				use: ['css-loader', 'postcss-loader']
+			}
+		]
+	}
+}
